Back-end/requirements.txt
flask
flask-jwt-extended
flask-cors
werkzeug
pymongo
python-dotenv